% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinPrecip.R
\name{basinPrecip}
\alias{basinPrecip}
\title{Accumulated precipitation}
\usage{
basinPrecip(location, start, end = Sys.time(), map = FALSE, raster_loc = NULL)
}
\arguments{
\item{location}{The location above which you wish to calculate precipitation. Specify either a WSC station ID (e.g. \code{"09AB004"}) or coordinates in signed decimal degrees in form latitude, longitude (\code{"60.1234 -139.1234"}; note the space, negative sign, and lack of comma). See details for more info if specifying coordinates.}

\item{start}{The start of the time period over which to accumulate precipitation. Use format \code{"yyyy-mm-dd hh:mm"} (character vector) in local time See details if requesting earlier than 30 days prior to now.}

\item{end}{The end of the time period over which to accumulate precipitation. Other details as per \code{start}}

\item{map}{Should a map be output to the console? See details for more info.}

\item{raster_loc}{The location where precipitation rasters are to be downloaded. Suggested use is to specify a repository where all such rasters are saved to speed processing time and reduce data usage. If using the default NULL, rasters will not persist beyond your current R session.}
}
\value{
The accumulated precipitation in mm of water above the WSC station or immediately surrounding the point specified in \code{location} and optionally, a map of precipitation printed to the console. In addition, the results can be assigned to a list.
}
\description{
Calculates accumulated precipitation above Water Survey of Canada stations or immediately surrounding a specified point. If possible, make sure to specify a location in raster_loc to speed things up!
}
\details{
If specifying \code{location} as coordinates rather than a WSC station, the numeric result will be of precipitation in the 2.5 km2 cell on which the point falls. Map output, if requested, will be for the smallest available WSC drainage polygon containing that point.

Raster images of precipitation are only kept by ECCC for 30 days. If specifying a \code{start} or \code{end} prior to that without having precipitation rasters available locally and specified with \code{raster_loc}, the start date/time will be adjusted and a warning given.

The \code{map} output, if requested, contains layers of streams, lakes, communities, major roads, and WSC stations to more easily situate yourself.

Rasters must be downloaded for each 6-hour period between \code{start} and \code{end}. It is therefore strongly suggested that you designate and use a folder for this purpose. \code{WRBtools::getHRDPA} is called for downloading and clipping rasters.
}
