% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levels_with_returns_mASL.R
\name{daily_level_data}
\alias{daily_level_data}
\title{Daily level data}
\usage{
daily_level_data(
  station_number,
  select_years = c(1950:2020),
  level_zoom = FALSE
)
}
\arguments{
\item{station_number}{The WSC station number for which you want data}

\item{select_years}{The year(s) for which you want data}

\item{level_zoom}{TRUE/FALSE, should high-res data be kept for zoomed-in plots?}
}
\value{
A list containing one element (a list of 4 with level data and historical data) if level_zoom is FALSE, or a list of 2 if level_zoom is TRUE (addition of high-res level data)
}
\description{
Daily level data
}
