% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECCCdata.R
\name{ECCCdata}
\alias{ECCCdata}
\title{Download and process ECCC flow or level data}
\usage{
ECCCdata(
  stations,
  level_flow = "Both",
  years = lubridate::year(Sys.Date()),
  recent_percentile = FALSE,
  filter = TRUE
)
}
\arguments{
\item{stations}{The stations for which you want to download data.}

\item{level_flow}{Do you want data for levels, flows, or both? Choose from "Level", "Flow", or "Both".}

\item{years}{The years for which you want easy-to-plot daily means. The resultant data.frame includes calculated To simplify plotting multiple years together, each daily mean data point is dated as if it was collected in the year 2022. Individual years are identified by the Year_Real column. Defaults to the current year.}

\item{recent_percentile}{Should the percent of historical max flows be calculated for recent data? Adds about 30 seconds per station.}

\item{filter}{Should the recent_data (if requested) be filtered to remove spikes? Adds about a minute of processing time per station.}
}
\value{
A list with an element for each station requested.
}
\description{
Neatly packages the output from utils_flow_data into a list with an element for each station requested. Output for each station are three data.frames: one of all-time historical data, one with the years requested in an easy to plot format, and one with the last 18 months of 5-minute data. Statistics are calculated for all data.frames.
}
