% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodReport.R
\name{floodReport}
\alias{floodReport}
\title{Flood reporting utility}
\usage{
floodReport(
  report_name = NULL,
  custom_report_stations = NULL,
  image_path = NULL,
  save_path = "choose",
  hydat_creds = "Renviron"
)
}
\arguments{
\item{report_name}{The name of the report you wish to generate. One of "Dawson", "Whitehorse", "Carmacks", "Ross River". Leave as default NULL if specifying a parameter under custom_report_stations.}

\item{custom_report_stations}{A user-specified list of stations for which to generate a report. Defaults to NULL to operate on the report_name parameter instead. Set to "choose" to navigate to the file using the GUI, or enter the file path directly as a character vector. The Excel document must be in .xlsx format with two columns: column 1 with the station name (as you want it to appear in the report), column 2 with the 7-character Station ID. The file must NOT contain a header.}

\item{image_path}{The path to the directory (folder) containing the images you wish to include. Default to NULL to not include any images. Set to "choose" to navigate to the folder, or enter the folder path directly as a character string. Some reports automatically include web-hosted images, these should not be included here.}

\item{save_path}{The path to the directory (folder) where the report should be saved. Default ("choose" )}

\item{hydat_creds}{The login credentials you wish to use for hydat. Defaults to your .Renviron file (so you can set and forget on a particular machine) but can also be specified as a character vector of two (as in c("username", "password")).}
}
\value{
A flood report containing flow and water level information in Microsoft Word format.
}
\description{
This function will allow for the generation of preset flood reports with minimal user inputs, and can generate reports based on a custom list of stations.  A Word document will be generated in which edits and additional images/information can be incorporated.
}
