% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrometric_Flow.R
\name{flow_data}
\alias{flow_data}
\title{Title}
\usage{
flow_data(
  station_number,
  extract_realtime = FALSE,
  select_years = 2020,
  csv_path = NA
)
}
\arguments{
\item{station_number}{}

\item{extract_realtime}{}

\item{select_years}{}

\item{csv_path}{}
}
\value{
A data.frame of timeseries data.
}
\description{
Title
}
