% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrometric_Flow.R
\name{flow_plot}
\alias{flow_plot}
\title{Plot flow data for a WSC station}
\usage{
flow_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  png_path = NA,
  historic = TRUE,
  log_scale = FALSE,
  percentile_plot = FALSE,
  colour_ramp = rainbow,
  line_colour_1 = "blue",
  line_colour_2 = "black",
  legend_position = "right",
  line_size = 0.75,
  point_size = 0.75,
  y_min = NA,
  y_max = NA
)
}
\arguments{
\item{station_number}{}

\item{complete_year}{}

\item{plot_years_df}{}

\item{dummy_year_df}{}

\item{png_path}{}

\item{historic}{}

\item{log_scale}{}

\item{percentile_plot}{}

\item{colour_ramp}{}

\item{line_colour_1}{}

\item{line_colour_2}{}

\item{legend_position}{}

\item{line_size}{}

\item{point_size}{}

\item{y_min}{}

\item{y_max}{}
}
\value{
A plot of flow volumnes for a WSC station.
}
\description{
Plot flow data for a WSC station
}
