% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levels_with_returns_mASL.R
\name{level_data}
\alias{level_data}
\title{Level data}
\usage{
level_data(
  station_number,
  extract_realtime = FALSE,
  select_years = c(1950:2020),
  csv_path = NA
)
}
\arguments{
\item{station_number}{The WSC station number for which you want data}

\item{extract_realtime}{Realtime data, TRUE or FALSE? TRUE requires login credentials for use with tidyhydat.ws.}

\item{select_years}{The year(s) for which you want data}

\item{csv_path}{Not used in this situation.}
}
\value{
A data.frame containing the timeseries requested.
}
\description{
For the purposes of this package, this function is simply a utility which facilitates the plotting of WSC data.
}
