% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrometric_Level_mASL.R,
%   R/Levels_with_returns.R
\name{level_plot}
\alias{level_plot}
\title{Plot WSC hydrometric data.}
\usage{
level_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  line_colour = "blue",
  legend_position = "right",
  line_size = 0.75,
  point_size = 0.75
)

level_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  line_colour = "blue",
  legend_position = "right",
  line_size = 0.75,
  point_size = 0.75
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{complete_year}{The year for which you want to plot.}

\item{plot_years_df}{Don't worry about this}

\item{dummy_year_df}{Also don't worry about this.}

\item{line_colour}{Self explanatory.}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}

\item{png_path}{Use this if you want to save the graph to a .png}

\item{historic}{No idea what this does}

\item{log_scale}{Log or regular scale?}

\item{percentile_plot}{Percentile or not?}

\item{colour_ramp}{What color do you feel like looking at?}

\item{line_colour_1}{Self explanatory.}

\item{line_colour_2}{Self explanatory.}

\item{y_min}{Self explanatory. Leave as NA to self-adjust.}

\item{y_max}{Self explanatory. Leave as NA to self-adjust.}
}
\value{
A plot for the station requested.

A plot for the station requested with return intervals, if they exist in the data.
}
\description{
For the purposes of this package, this function is simply a utility which facilitates the plotting of WSC data.

Function to plot water levels while including return intervals, where those exist and are specified in this package's data file named return_periods.
}
