% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelPlot.R
\name{levelPlot}
\alias{levelPlot}
\title{Level plots of WSC data}
\usage{
levelPlot(
  station,
  years,
  title = TRUE,
  zoom = FALSE,
  zoom_days = 30,
  filter = FALSE,
  forecast = NULL,
  returns = "auto",
  force_CGVD28 = FALSE,
  save_path = "none"
)
}
\arguments{
\item{station}{The WSC station for which you wish to generate a plot as a character vector of length 1.}

\item{years}{The year(s) you wish to plot. Maximum of 10 years specified in a numeric vector.}

\item{title}{Do you want a title added to the plot? TRUE/FALSE.}

\item{zoom}{TRUE/FALSE. If TRUE, the plot x axis (dates) will be truncated to the number of days prior to today specified in zoom_days.}

\item{zoom_days}{Number from 1 to 365. Not used unless zoom=TRUE.}

\item{filter}{TRUE/FALSE. Should 5-minute data be filtered to remove spikes? Adds about a minute per graph.}

\item{forecast}{Not currently in use; will eventually work similarly to forecast in flowPlot.}

\item{returns}{Should level returns be added? You have the option of using pre-determined level returns only (option "table"), auto-calculated values with no human verification (option "calculated", calculated on-the-fly using all data available from March to September, up to the current date), or "auto" (priority to pre-determined levels, fallback to auto-calculated levels), or none (option "none"). Defaults to "auto".}

\item{force_CGVD28}{For stations with a datum, should CGVD28 be used even if there is a more recent datum? Default FALSE will use the most recent available.}

\item{save_path}{Default is "none", and the graph will be visible in RStudio and can be assigned to an object. Option "choose" brings up the File Explorer for you to choose where to save the file, or you can also specify a save path directly.}
}
\value{
A .png file of the plot requested (if a save path has been selected), plus the plot displayed in RStudio. Assign the function to a variable to also get a plot in your global environment.
}
\description{
Generates plots of water levels from Water Survey of Canada stations, with up to 10 years specified by the user.
}
\details{
To generate zoomed-in plots with real-time data you MUST have your hydat credentials loaded into your .Renviron profile as values pairs of WS_USRNM=”your_username” and WS_PWD=”your_password”.
}
