% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPlot.R
\name{multiPlot}
\alias{multiPlot}
\title{Plot multiple WSC stations on one graph.}
\usage{
multiPlot(stations, type, days = 30, title = NULL, save_path = "choose")
}
\arguments{
\item{stations}{The WSC stations for which you wish to generate a plot. The first one listed should be the primary station, and usually the one with the greatest flow or slowest level response. If plotting flows, the left y-axis will be for this station while subsequent stations will plot on the secondary y-axis.}

\item{type}{"Level" or "Flow"?}

\item{days}{Number from 2 to 730 representing the number of days to plot.}

\item{title}{Do you want a title added to the plot? Leave as NULL for no title, otherwise enter it here as a character string.}

\item{save_path}{Where you wish to save the plot. Default is "choose" which brings up the File Explorer for you to choose.}
}
\value{
A .png file of the plot requested. Assign the function to a variable to also get a plot in your global environment.
}
\description{
Generates overlapping plots of water levels or flows from Water Survey of Canada stations, plotting up to 10 stations together at once. Only plots data as far back as 18 months from today's date. Warning: this function can take a long time to execute!
}
\details{
Only points are plotted with no connecting lines: this ensures that gaps or excessive noise in the data is clearly visible.

As with other package functions, you MUST have your hydat credentials loaded into your .Renviron profile as values pairs of WS_USRNM=”your_username” and WS_PWD=”your_password”.
}
