% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{return_periods}
\alias{return_periods}
\title{Return intervals for water levels at several stations in Yukon}
\format{
A data frame containing information on return intervals of water level (2 to 2000 years) for 14 stations in Yukon.
}
\usage{
return_periods
}
\description{
Return intervals for water levels at several stations in Yukon
}
\keyword{datasets}
