% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_stns}
\alias{spatial_stns}
\title{Table with WSC, ECCC weather reporting stations, and ECCC meteograms coordinates and basic information}
\format{
A data frame
}
\usage{
spatial_stns
}
\description{
Table with WSC, ECCC weather reporting stations, and ECCC meteograms coordinates and basic information
}
\keyword{datasets}
