% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMESH.R
\name{splitMESH}
\alias{splitMESH}
\title{Split MESH pdf output into individual pdfs}
\usage{
splitMESH(file = "choose", output_folder = "choose")
}
\arguments{
\item{file}{The pdf file you wish to split. Defaults "choose" lets you interactively choose a file.}

\item{output_folder}{The folder in which the new pdfs will be saved. Default "choose" lets you interactively choose a folder.}
}
\value{
A pdf for each station in the output pdf.
}
\description{
This utility splits the standard MESH pdf output. Predicated on each page of the pdf containing one station (no more, no less). Outputs with stations taking up more than one page or with two stations per page will not work in this current state.
}
