% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_station.R
\name{SWE_station}
\alias{SWE_station}
\title{Calculates snow survey stats for stations}
\usage{
SWE_station(file_loc, year, month, threshold = 7, csv = FALSE)
}
\arguments{
\item{file_loc}{File location of the Factors table}

\item{year}{The year of interest. The stats will be calculated based on all years prior to 'year'.}

\item{month}{The month of interest. Options are 3, 4 and 5 for March, April and May, respectively. Historical stats are given for the first day of this month.}

\item{threshold}{A number between 1 and 10 giving the threshold below which the SWE for that basin and year are ignored. These numbers represent the sum of the factors of the stations for a basin which are not missing data for that year. 10 means that the swe values calculated from less than all the stations of that basin are ignored. 1 means that only the swe calculated from less than 1 out of 10 are ignored.}

\item{csv}{TRUE or FALSE. If TRUE, a csv will be created.}
}
\value{
A table and a csv file (if csv = TRUE) with the current SWE, historical median, the swe relative to the median (swe / swe_median), historical maximum, historical minimum, and year of maximum and minimum for each basin.
}
\description{
#' @description
`r lifecycle::badge('stable')`
# The purpose of this script is to automate the SWE calculator for YG snow survey data. It has been modified from Ellen Ward's code from 2020-04-16.
}
