% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularReport.R
\name{tabularReport}
\alias{tabularReport}
\title{Tabular output of hydrometric data}
\usage{
tabularReport(
  database = NULL,
  water_locations = "default",
  past = 31,
  level = TRUE,
  flow = TRUE,
  pillows = TRUE,
  save_path = "choose"
)
}
\arguments{
\item{database}{If using a local database created using WRBdatabase package, specify its path here. Leave NULL to download from the Water Survey of Canada and/or Aquarius instead. See details.}

\item{water_locations}{List of hydrometric locations to include in the report, as a character vector. Default is a pre-determined list of locations across the territory.}

\item{past}{The number of days in the past for which you want data. Will be rounded to yield table columns covering up to 1 month, in 1-week increments. At minimum a 1-week column will be populated, at maximum a 1-month column and 1, 2, 3 week columns.}

\item{level}{Level table, TRUE/FALSE}

\item{flow}{Flow table, TRUE/FALSE}

\item{pillows}{Snow pillow table, TRUE/FALSE}

\item{save_path}{}
}
\value{
A Word document containing the report.
}
\description{
Creates a report of hydrometric and snow pack conditions, each table separated by a page break, landscape format. List of hydrometric stations/locations can be user-defined, snow pillow locations are present to include all active pillows as of 2023-01-31.
Database connection should be prefered over direct pulls for speed, with nearly no lag in data availability between the two. Connection is established using WRBtools::hydroConnect, so ensure that WRBtools is up to date if the database type has changed.
}
