% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfutils_flow.R, R/utils_flow.R
\name{utils_daily_flow_plot}
\alias{utils_daily_flow_plot}
\title{Plot flow data for a WSC station}
\usage{
utils_daily_flow_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  colours = c("blue", "darkorange", "darkorchid3", "cyan2", "firebrick3",
    "aquamarine4", "gold1", "chartreuse1", "black"),
  legend_position = "right",
  line_size = 1,
  point_size = 0.75
)

utils_daily_flow_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  colours = c("blue", "darkorange", "darkorchid3", "cyan2", "firebrick3",
    "aquamarine4", "gold1", "chartreuse1", "black"),
  legend_position = "right",
  line_size = 1,
  point_size = 0.75
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{complete_year}{The current year, normally output from daily_flow_data.}

\item{plot_years_df}{data.frame containing plotting data for all years selected, normally output from daily_flow_data}

\item{dummy_year_df}{Output from daily_flow_data}

\item{colours}{Colour for the lines and points}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}

\item{png_path}{}

\item{historic}{}

\item{log_scale}{}

\item{percentile_plot}{}

\item{colour_ramp}{}

\item{line_colour_1}{}

\item{line_colour_2}{}

\item{y_min}{}

\item{y_max}{}
}
\value{
A plot of flow volumes for a WSC station.

A plot of flow volumes for a WSC station.
}
\description{
This utility function is designed to take the output of the utils_flow_data function. If you're looking for a plot, use the flowPlot function instead.
}
