% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_zoom_fcast.R
\name{utils_fcast_level_plot}
\alias{utils_fcast_level_plot}
\title{Plot WSC hydrometric level data for a set number of days using 5 minute data points for the current year as well as MESH and CLEVER forecast data.}
\usage{
utils_fcast_level_plot(
  station_number,
  level_years,
  zoom_data,
  zoom_days = 30,
  colours = c("blue", "black", "darkorchid3", "cyan2", "firebrick3", "aquamarine4",
    "gold1", "chartreuse1", "darkorange", "lightsalmon"),
  legend_position = "right",
  line_size = 1,
  point_size = 0.75
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{level_years}{A data.frame of plotting data}

\item{zoom_data}{The data frame of zoomed-in data.}

\item{zoom_days}{The number of days to plot, counting back from the current date.}

\item{colours}{Colour of the lines/points.}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}
}
\value{
A plot for the station requested and for the duration requested.
}
\description{
This utility function is designed to work with objects generated in the condition reports markdown file. If you're looking for a plot, use the levelPlot function instead.
}
