% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_level.R
\name{utils_level_data}
\alias{utils_level_data}
\title{Download level data}
\usage{
utils_level_data(station_number, select_years, level_zoom = TRUE)
}
\arguments{
\item{station_number}{The WSC station number for which you want data.}

\item{select_years}{The year(s) for which you want data}

\item{level_zoom}{TRUE/FALSE, should high-res data be kept for zoomed-in plots?}
}
\value{
A list containing one element (a list of 4 with level data and historical data) if level_zoom is FALSE, or a list of 2 if level_zoom is TRUE (addition of high-res level data)
}
\description{
Utility function to download water level data from WSC online databases. If you are looking for data in an easy to use format please use levelData function instead.
}
