% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_zoom_flow_plot.R
\name{utils_zoom_flow_plot}
\alias{utils_zoom_flow_plot}
\title{Plot WSC hydrometric flow data for a set number of days using 5 minute data points for the current year}
\usage{
utils_zoom_flow_plot(
  station_number,
  flow_years,
  zoom_data,
  complete_df,
  zoom_days = 30,
  colours = c("blue", "black", "darkorchid3", "cyan2", "firebrick3", "aquamarine4",
    "gold1", "chartreuse1", "darkorange", "lightsalmon"),
  legend_position = "right",
  line_size = 1,
  point_size = 0.75,
  returns = "none"
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{flow_years}{A data.frame of plotting data}

\item{zoom_data}{The data frame of zoomed-in data.}

\item{complete_df}{data.frame containing historical data from the start of records to the last year to be plotted.}

\item{zoom_days}{The number of days to plot, counting back from the current date.}

\item{colours}{Colour of the lines/points.}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}

\item{returns}{Should flow returns be calculated, plotted, and added to the flows table? You have the option of using pre-determined flows only (option "table"), auto-calculated values with no human verification (option "auto", calculated on-the-fly using all data available from March to September, up to the current date), both (with priority to pre-determined flows), or none (option "none"). Defaults to "none".}
}
\value{
A plot for the station requested and for the duration requested.
}
\description{
This utility function is designed to take the output of the utils_flow_data function. If you're looking for a plot, use the flowPlot function instead.
}
