% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_level.R
\name{utils_zoom_level_plot}
\alias{utils_zoom_level_plot}
\title{Plot WSC hydrometric level data for a set number of days using 5 minute data points for the current year.}
\usage{
utils_zoom_level_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  zoom_data,
  zoom_days = 30,
  colours = c("blue", "darkorange", "darkorchid3", "cyan2", "firebrick3",
    "aquamarine4", "gold1", "chartreuse1", "black", "lightsalmon"),
  legend_position = "right",
  line_size = 1,
  point_size = 0.75
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{complete_year}{The year for which you want to plot.}

\item{plot_years_df}{A data.frame of plotting data}

\item{dummy_year_df}{A data.frame containing only the min, max, and quartiles.}

\item{zoom_data}{The data frame of zoomed-in data.}

\item{zoom_days}{The number of days to plot, counting back from the current date.}

\item{colours}{Colour of the lines/points.}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}
}
\value{
A plot for the station requested and for the duration requested.
}
\description{
This utility function is designed to take the output of the utils_level_data function. If you're looking for a plot, use the levelPlot function instead.
}
