% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levels_with_returns_mASL.R
\name{zoom_level_plot}
\alias{zoom_level_plot}
\title{Zoomed-in plot of WSC data.}
\usage{
zoom_level_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  zoom_data,
  zoom_days = 30,
  colours = c("blue", "green4", "darkorchid3", "orange1", "aquamarine4", "cyan2"),
  legend_position = "right",
  line_size = 1,
  point_size = 1
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{complete_year}{The year for which you want to plot.}

\item{plot_years_df}{Don't worry about this}

\item{dummy_year_df}{Also don't worry about this.}

\item{zoom_data}{The data frame of zoomed-in data.}

\item{zoom_days}{The number of days to plot, counting back from the current date.}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}

\item{colour}{Colour of the lines/points.}
}
\value{
A plot for the station requested and for the duration requested.
}
\description{
Zoomed-in plot of WSC data.
}
