% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levels_with_returns_mASL.R
\name{daily_level_data}
\alias{daily_level_data}
\title{Daily level data}
\usage{
daily_level_data(
  station_number,
  extract_realtime = FALSE,
  select_years = c(1950:2020)
)
}
\arguments{
\item{station_number}{The WSC station number for which you want data}

\item{extract_realtime}{Realtime data, TRUE or FALSE? TRUE requires login credentials for use with tidyhydat.ws.}

\item{select_years}{The year(s) for which you want data}
}
\value{
A data.frame containing the timeseries requested.
}
\description{
Daily level data
}
