% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levels_with_returns_mASL.R
\name{daily_level_plot}
\alias{daily_level_plot}
\title{Plot WSC hydrometric data.}
\usage{
daily_level_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  line_colour = "blue",
  legend_position = "right",
  line_size = 0.5,
  point_size = 1.5
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{complete_year}{The year for which you want to plot.}

\item{plot_years_df}{Don't worry about this}

\item{dummy_year_df}{Also don't worry about this.}

\item{line_colour}{Self explanatory.}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}
}
\value{
A plot for the station requested with return intervals, if it exists in the data file data$return_periods.
}
\description{
Function to plot water levels while including return intervals, where those exist and are specified in this package's data file named data$return_periods.
}
