% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_ALL_reports.R, R/Generate_report.R
\name{floodReport}
\alias{floodReport}
\title{Flood reporting utility}
\usage{
floodReport(
  report_name,
  image_path = NULL,
  save_path = "choose",
  hydat_creds = "Renviron"
)

floodReport(
  report_name,
  image_path = NULL,
  save_path = "choose",
  hydat_creds = "Renviron"
)
}
\arguments{
\item{report_name}{The name of the report you wish to generate. One of "Dawson", "Whitehorse", "Carmacks", "Ross River". Leave as default NULL if specifying a parameter under custom_report_stations.}

\item{image_path}{The path to the directory (folder) containing the images you wish to include. Default to NULL to not include any images. Set to "choose" to navigate to the folder, or enter the folder path directly as a character string. Some reports automatically include web-hosted images, these should not be included here.}

\item{save_path}{The path to the directory (folder) where the report should be saved. Default ("choose" )}

\item{hydat_creds}{The login credentials you wish to use for hydat. Defaults to your .Renviron file (so you can set and forget on a particular machine) but can also be specified as a character vector of two (as in c("username", "password")).}

\item{custom_report_stations}{A user-specified list of stations for which to generate a report. Defaults to NULL to operate on the report_name parameter instead. Input must be a character vector of all station IDs. Reminder: you can create a character vector from a column of a data.frame, and you can reference an environment object instead of typing in the vector!}

\item{report_type}{What do you want your report to contain? Choose from "Level", "Flow", or "Both." Defaults to Both.}

\item{level_zoom}{Do you want a zoomed-in plot for level? Choose from TRUE or FALSE. Defaults to FALSE.}

\item{zoom_scale}{The number of days to plot for zoomed in level plots. Defaults to 40, but not used unless level_zoom is set to TRUE.}
}
\value{
A flood report containing flow and water level information in Microsoft Word format.

A flood report containing flow and water level information in Microsoft Word format.
}
\description{
This function will allow for the generation of preset flood reports with minimal user inputs, and can generate reports based on a custom list of stations.  A Word document will be generated in which edits and additional images/information can be incorporated.

This function will allow for the generation of preset flood reports with minimal user inputs, and can generate reports based on a custom list of stations.  A Word document will be generated in which edits and additional images/information can be incorporated.
}
\details{
Note that you MUST have your hydat credentials loaded into your .Renviron profile as values pairs of WS_USRNM=”your_username” and WS_PWD=”your_password”. Refer to the R and GitHub for the WRB word document for more information.
}
