% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrometric_Flow.R
\name{flow_data}
\alias{flow_data}
\title{Title}
\usage{
flow_data(
  station_number,
  extract_realtime = FALSE,
  select_years = 2020,
  csv_path = NA
)
}
\arguments{
\item{station_number}{The WSC station for which you want to fecth data.}

\item{extract_realtime}{Should real-time data be extracted, or only historical data?}

\item{select_years}{The years for which data should be plotted.}

\item{csv_path}{}
}
\value{
A data.frame of timeseries data.
}
\description{
Title
}
