% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowPlot.R
\name{flowPlot}
\alias{flowPlot}
\title{Flow plots of WSC data}
\usage{
flowPlot(station, years, zoom = FALSE, zoom_days = 30, save_path = "choose")
}
\arguments{
\item{station}{The WSC station for which you wish to generate a plot.}

\item{years}{The year(s) you wish to plot. Maximum of 10 years specified in a vector.}

\item{zoom}{TRUE/FALSE. If TRUE, the plot x axis (dates) will be truncated to the number of days prior to today specified in zoom_days.}

\item{zoom_days}{Number from 1 to 365. Not used unless zoom=TRUE.}

\item{save_path}{Where you wish to save the plot. Default is "choose" which brings up the File Explorer for you to choose.}
}
\value{
A .png file of the plot requested, plus the plot displayed in RStudio. Assign the function to a variable to also get a plot in your global environment.
}
\description{
Generates plots of water flows from Water Survey of Canada stations, with up to 10 years specified by the user.
}
\details{
To generate zoomed-in plots with real-time data you MUST have your hydat credentials loaded into your .Renviron profile as values pairs of WS_USRNM=”your_username” and WS_PWD=”your_password”.

You must also manually install the dependent package "tidyhydat.ws" as it lives on a github repository. Use install.packages('tidyhydat.ws', repos='https://bcgov.github.io/drat/')
}
