% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flow.R
\name{utils_daily_flow_plot}
\alias{utils_daily_flow_plot}
\title{Plot WSC hydrometric flow data for the whole year using daily means.}
\usage{
utils_daily_flow_plot(
  station_number,
  flow_years,
  colours = c("blue", "black", "darkorchid3", "cyan2", "firebrick3", "aquamarine4",
    "gold1", "chartreuse1", "darkorange", "lightsalmon"),
  legend_position = "right",
  line_size = 1,
  point_size = 0.75
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{flow_years}{data.frame containing plotting data for all years selected, normally output from daily_flow_data}

\item{colours}{Colour for the lines and points}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}
}
\value{
A plot of flow volumes for a WSC station.
}
\description{
This utility function is designed to take the output of the utils_flow_data function. If you're looking for a plot, use the flowPlot function instead.
}
