% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_daily_flow_plot.R
\name{utils_daily_flow_plot}
\alias{utils_daily_flow_plot}
\title{Plot WSC hydrometric flow data for the whole year using daily means.}
\usage{
utils_daily_flow_plot(
  station_number,
  flow_years,
  complete_df,
  colours = c("blue", "black", "darkorchid3", "cyan2", "firebrick3", "aquamarine4",
    "gold1", "chartreuse1", "darkorange", "lightsalmon"),
  line_size = 1,
  point_size = 0.75,
  returns = "auto"
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{flow_years}{data.frame containing plotting data for all years selected, normally output from daily_flow_data}

\item{complete_df}{data.frame containing historical data from the start of records to the last year to be plotted.}

\item{colours}{Colour for the lines and points}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}

\item{returns}{Should flow returns be plotted? You have the option of using pre-determined flow returns only (option "table"), auto-calculated values with no human verification (option "auto", calculated on-the-fly using all data available from March to September, up to the current date), "both" (with priority to pre-determined flows), or none (option "none"). Defaults to "both".}
}
\value{
A plot of flow volumes for a WSC station.
}
\description{
This utility function is designed to take the output of the utils_flow_data function. If you're looking for a plot, use the flowPlot function instead.
}
