% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_level.R
\name{utils_daily_level_plot}
\alias{utils_daily_level_plot}
\title{Plot WSC hydrometric level data for the whole year using daily means.}
\usage{
utils_daily_level_plot(
  station_number,
  complete_year,
  plot_years_df,
  dummy_year_df,
  colours = c("blue", "black", "darkorchid3", "cyan2", "firebrick3", "aquamarine4",
    "gold1", "chartreuse1", "darkorange", "lightsalmon"),
  legend_position = "right",
  line_size = 1,
  point_size = 0.75
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{complete_year}{The current year, normally output from daily_level_data.}

\item{plot_years_df}{data.frame containing plotting data for all years selected, normally output from daily_level_data}

\item{dummy_year_df}{Output from daily_level}

\item{colours}{Colour for the lines and points}

\item{legend_position}{Self explanatory.}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}
}
\value{
A plot for the station requested with return intervals, if it exists in the data file data$return_periods.
}
\description{
This utility function is designed to take the output of the utils_level_data function. If you're looking for a plot, use the levelPlot function instead.
}
