% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_daily_level_plot.R
\name{utils_daily_level_plot}
\alias{utils_daily_level_plot}
\title{Plot WSC hydrometric level data for the whole year using daily means.}
\usage{
utils_daily_level_plot(
  station_number,
  level_years,
  complete_df,
  colours = c("blue", "black", "darkorchid3", "cyan2", "firebrick3", "aquamarine4",
    "gold1", "chartreuse1", "darkorange", "lightsalmon"),
  line_size = 1,
  point_size = 0.75,
  returns = "auto",
  force_CGVD28 = FALSE
)
}
\arguments{
\item{station_number}{The station for which you want to plot data.}

\item{level_years}{data.frame containing plotting data for all years selected, normally output from daily_level_data}

\item{complete_df}{data.frame containing historical data from the start of records to the last year to be plotted.}

\item{colours}{Colour for the lines and points}

\item{line_size}{Self explanatory.}

\item{point_size}{Self explanatory.}

\item{returns}{Should level returns be added? You have the option of using pre-determined level returns only (option "table"), auto-calculated values with no human verification (option "calculated", calculated on-the-fly using all data available from March to September, up to the current date), or "auto" (priority to pre-determined levels, fallback to auto-calculated levels), or none (option "none"). Defaults to "auto".}

\item{force_CGVD28}{For stations with a datum, should CGVD28 be used even if there is a more recent datum?}
}
\value{
A plot for the station requested with return intervals, if it exists in the data file data$level_returns.
}
\description{
This utility function is designed to take the output of the utils_level_data function. If you're looking for a plot, use the levelPlot function instead.
}
