% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flow.R
\name{utils_flow_data}
\alias{utils_flow_data}
\title{Download flow data}
\usage{
utils_flow_data(station_number, select_years, flow_zoom = TRUE, filter = FALSE)
}
\arguments{
\item{station_number}{The WSC station for which you want data.}

\item{select_years}{The years for which you want data.}

\item{flow_zoom}{TRUE/FALSE, should high-res data be kept for zoomed-in plots?}

\item{filter}{TRUE/FALSE, should recent data be filtered to remove spikes? Adds about a minute for each station.}
}
\value{
A list containing three elements: a data.frame of all historical data, a data.frame containing data for the years requested with min, max, and percentiles calculated, and a data.frame containing 5-minute data for the past 18 months.
}
\description{
Utility function to download water flow data from WSC online databases. If you are looking for data in an easy to use format please use flowData function instead.
}
