% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfutils_flow.R, R/utils_flow.R
\name{utils_flow_data}
\alias{utils_flow_data}
\title{Download flow data}
\usage{
utils_flow_data(station_number, select_years, flow_zoom = TRUE)

utils_flow_data(station_number, select_years, flow_zoom = TRUE)
}
\arguments{
\item{station_number}{The WSC station for which you want data.}

\item{select_years}{The years for which you want data.}

\item{flow_zoom}{TRUE/FALSE, should high-res data be kept for zoomed-in plots?}
}
\value{
A list containing one element (a list of 4 with level data and historical data) if level_zoom is FALSE, or a list of 2 if level_zoom is TRUE (addition of high-res level data)

A list containing one element (a list of 4 with flow data and historical data) if flow_zoom is FALSE, or a list of 2 if flow_zoom is TRUE (addition of high-res flow data)
}
\description{
Utility function to download water flow data from WSC online databases. If you are looking for data in an easy to use format please use flowData function instead.

Utility function to download water flow data from WSC online databases. If you are looking for data in an easy to use format please use flowData function instead.
}
