% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_level_data.R
\name{utils_level_data}
\alias{utils_level_data}
\title{Download level data}
\usage{
utils_level_data(
  station_number,
  select_years,
  high_res = TRUE,
  filter = TRUE,
  recent_prctile = FALSE,
  rate = FALSE,
  rate_days = "all"
)
}
\arguments{
\item{station_number}{The WSC station number for which you want data.}

\item{select_years}{The year(s) for which you want data.}

\item{high_res}{TRUE/FALSE, should high-res data be kept for zoomed-in plots? Default FALSE.}

\item{filter}{TRUE/FALSE, should recent data be filtered to remove spikes? Adds about a minute for each station, default FALSE.}

\item{recent_prctile}{TRUE/FALSE, should the recent (5 minute) data have a percent of maximum historical levels calculated? Adds about 30 seconds, default FALSE.}

\item{rate}{TRUE/FALSE, should the difference from one data point compared to the previous data point be calculated into a new column? Adds about 1.5 minutes for all data points, default FALSE. If high_res == TRUE, rate is only calculated for the data.frame containing daily means. This data will likely be noisy, a rolling mean might be better.}

\item{rate_days}{Number days for which to calculate a rate of change, applied only to high-resolution data (historical daily means data is quick to calculate and all days are automatically calculated). Defaults to "all" which calculates rates for all 18 months of past high-resolution level data; specify a smaller number of days as an integer to lessen processing time.}
}
\value{
A list containing three elements: a data.frame of all historical data, a data.frame containing data for the years requested with min, max, and percentiles calculated, and a data.frame containing high-resolution data if the requested years encompass the previous 18 months. To facilitate plotting, the data.frame with requested years (list element 2) has a column of "fake" dates where each year of data has dates as if they were in the most recent year requested; the true year is contained in the Year_Real column.
}
\description{
Utility function to download water level data from WSC online databases. If you are looking for data in an easy to use format please use WRBfloods::levelData function instead.
}
