% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterData.R
\name{waterData}
\alias{waterData}
\title{Return flow or level data}
\usage{
waterData(
  stations,
  level_flow = "Both",
  years = lubridate::year(Sys.Date()),
  recent_percentile = FALSE,
  filter = TRUE,
  rate = FALSE,
  rate_days = "all",
  force_CGVD28 = FALSE
)
}
\arguments{
\item{stations}{The stations for which you want to download data. If any of these do not exist in the database (and if you have specified a path for the database) then raw information will be downloaded from the WSC.}

\item{level_flow}{Do you want data for levels, flows, or both? Choose from "Level", "Flow", or "Both". Levels will be in the most recent datum available.}

\item{years}{The years for which you want easy-to-plot daily means. The resultant data.frame includes calculated To simplify plotting multiple years together, each daily mean data point is dated as if it was collected in the year 2022. Individual years are identified by the Year_Real column. Defaults to the current year.}

\item{recent_percentile}{Should the percent of historical max flows be calculated for recent data? Adds about 30 seconds per station.}

\item{filter}{Should the recent_data (if requested) be filtered to remove spikes? Adds about a minute of processing time per station.}

\item{rate}{TRUE/FALSE, should the difference from one data point compared to the previous data point be calculated into a new column? Adds about 1.5 minutes for all data points, default FALSE. This data will likely be noisy.}

\item{rate_days}{Number days for which to calculate a rate of change, applied only to high-resolution data recent data (historical daily means data is quick to calculate). Defaults to "all" which calculates rates for all 18 months of past high-resolution level data; specify a smaller number of days as an integer to lessen processing time.}

\item{force_CGVD28}{For stations with a datum and when levels information is requested, should CGVD28 be used even if there is a more recent datum?}
}
\value{
A list with an element for each station requested, each element consisting of three data.frames: one of all historical daily data, one with the years requested, and one with realtime information for the years requested.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Gets level or flow data from WSC locations directly from the Water Survey of Canada. Important: it's usually preferable to get flow/level data from the local database created with WRBdatabase in most cases, using \code{\link[WRBtools:DB_get_ts]{WRBtools::DB_get_ts()}}.

Neatly packages the output from two hidden utility functions into a list with an element for each station requested. Outputs for each station are three data.frames: one of all-time historical data, one with the years requested in an easy to plot format, and one with the last 18 months of 5-minute data. Statistics are calculated for all data.frames.
}
