% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRBfloods-package.R
\docType{package}
\name{WRBfloods-package}
\alias{WRBfloods}
\alias{WRBfloods-package}
\title{WRBfloods: WRB's flood report functions}
\description{
This package exists to facilitate the generation of flood reports by the Yukon Water Resources Branch.
}
\author{
\strong{Maintainer}: Ghislain de Laplante \email{ghislaindel@gmail.com} (\href{https://orcid.org/0000-0002-5093-9185}{ORCID})

}
\keyword{internal}
